requirements.txt
streamlit
Pillow
PyMuPDF
groq